#amazon #coding
Given an array of integers, we sum neighboring elements like pascal's triangle until only two integers remain and take those to be the output. During the process, if the sum of two elements is greater than or equal to 10, we only take the digit place to the next level. Sorry if the description isn't clear enough, I included a picture to demonstrate the idea.